finish up fast json
