+++ polars/polars-io/src/ndjson_core/mod.rs
@@ -1,2 +1,2 @@
 pub(crate) mod buffer;
-pub mod json;
+pub mod ndjson;
